requirements.txt
asgiref==3.2.7
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
Django==3.0.4
django-crispy-forms==1.9.0
django-mysql==3.3.0
django-qr-code==1.1.0
django-qrcode==0.3
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
gunicorn==20.0.4
mysqlclient==1.4.6
Pillow==7.1.2
pycparser==2.20
pypng==0.0.20
PyQRCode==1.2.1
pytz==2019.3
qrcode==5.1
six==1.14.0
sqlparse==0.3.1